refactor: message event logging util funcs
Remove `LoggingModule.logMessageEdit` and
`LoggingModule.logMessageDelete`. Move said functionality to
their respective event handlers.
